Extend providers to support custom headers
We have a provider that requires custom headers for authentication
rather than path details, so I've extended the plugin to support parsing
that out of the Caddyfile and providing the headers in the outgoing
request after the provider is selected.

This also sets a framework for additional provider-specific details. When we
need to have provider weights, providers that only implement specific methods,
etc. this will be a good place to put it.