Implement provider tiering
This adds an optional "priority" field for providers. Traffic will
be routed to a server from a healthy provider with the lowest priority
number (0 first, then 1, then 2). The priority for a provider must not
be higher than the number of providers in a tier. If no healthy providers
are available after reviewing all priority levels, all servers will be
balanced.