Implement rudimentary healthchecks
Caddy's native healthchecks don't work with dynamic upstreams,
so we're implementing a very basic one for now. It makes an RPC request
to the target host every 5 seconds. 2 consecutive failures will mark
it as unhealthy, 2 consecutive successes will mark it as healthy.

The RPC method for checking defaults to "eth_blockNumber", but can
be configured differently if needed. At some point we'll need to
figure out health checks for non-rpc services, but there are other
factors to consider there.