NFT support
This implements the provider-side of an NFT authentication protocol.

If a signer is not in the whitelist, it checks the NFT specified
in the resources section of the SIWE claim, and if the NFT corresponds
to this provider and is owned by the signer, the session token will
be issued.