Make auth whitelist case insensitive
CompareNodes gave us a key for the whitelist in an un-checksummed
form, so that's how it went into the whitelist. When we recovered
the address rom the signature, it came back in checksummed form, which
did not match.

This change shifts everything for the whitelist to lowercase, both
when reading in addresses and when recovering addresses.