Enable naive method-based routing
This implements a very simple version of method based routing. Networks
can specify a list of `routed_methods`, and providers can specify a list
of `methods`.

If the network specifies a list of `routed_methods`, any request that has
a method in the `routed_methods` list will only be eligible to be routed
to providers that specify the same method in `methods`.

This does very little to ensure consistency. It will make sure that method
routed requests with same Din-Session-Id will go to the same provider for
the same method, but that may be a different provider than non-method-routed
requests, so they may be on different blocks.