added support for additional block jump check support (#93)
* Enable naive method-based routing

This implements a very simple version of method based routing. Networks
can specify a list of `routed_methods`, and providers can specify a list
of `methods`.

If the network specifies a list of `routed_methods`, any request that has
a method in the `routed_methods` list will only be eligible to be routed
to providers that specify the same method in `methods`.

This does very little to ensure consistency. It will make sure that method
routed requests with same Din-Session-Id will go to the same provider for
the same method, but that may be a different provider than non-method-routed
requests, so they may be on different blocks.

* added support for additional block jump check support

* Documentation and logging updates based on PR feedback

* Fix filtered method check in logs

* Feat/support multiple endpoints with same host + docs (#92)

* support env and blocknumber in prometheus

* fixed test

* update health checks docs

* add health check status

* add reg sync doc

* added additional logging and status cleanup

* added support for unique host as middleware provider key

* moved where the network sepcific routing docs are found

* updated main readme with new docs

* added onchain word

* added support for additional block jump check support

* updated docs to include block jump updates

---------

Co-authored-by: Austin Roberts <austin.roberts@rivet.cloud>
Co-authored-by: Austin Roberts <git@ausiv.com>