Fix method filtering routing bug
The MethodFilter sets the list of available providers on the replacer
object, and to ensure this is honored that needs to be used by GetUpstreams.

I'm not sure why GetUpstreams was changed during the refactor to rely on
the network object. I suppose it's possible there's some situation where
it's coming back empty, and using networkConfig.Providers seemed more reliable,
but for the method filter to work correctly we need to use the same map
from the method filter. To mitigate any gaps in my own understanding,
I've implemented a failover so that if the providers retrieved from the
replacer is nil, we can still populate from the networkConfig.