Translate prometheus api keys
This update removes private API keys from prometheus data, replacing
them with non-sensitive data.

If allows for a map of known users, and will send on a name that corresponds
to the API key when available.

Since we're not currently trying to block unknown users, but collect data
about unknown users, if a key isn't in the map we'll send signoz a 10
character salted hash of the API key, so we can issue keys and identify
when they're being used without having to redeploy caddy.