Initial implementation of provider side NFT authentication
This is an initial (but at this point untested) implementation of
using Superfluid flows in conjunction with DIN NFTs.

It keeps track of the RPS and Superfluid flows of all issued sessions,
and when new sessions are requested it checks the requested RPS
against the required and available flow rates, and issues the session
token if there is enough payment flowing to cover the cost of the
session.

At this point there is a lot remaining for this to be complete, but most
of those items have been ticketed separately. Some examples are:

* RPS limits are not enforced
* The client library doesn't specify rps
* The client library cannot allocate flow
* The session tracker is currently in-memory only, with no shared state across instances

These items will be addressed in the future.